requirements.txt
pygame
numpy
svg.path